HSFM.FileSystem.FileType

Description

This module provides a data type for representing directories/files in a well-typed and convenient way. This is useful to gather and save information about a file, so the information can be easily processed in e.g. a GUI.

However, it's not meant to be used to interact with low-level functions that copy files etc, since there's no guarantee that the in-memory representation of the type still matches what is happening on filesystem level.

If you interact with low-level libraries, you must not pattern match on the `File a` type. Instead, you should only use the saved path and make no assumptions about the file the path might or might not point to.

Documentation

data File a

The String in the path field is always a full path. The free type variable is used in the File/Dir constructor and can hold Handles, Strings representing a file's contents or anything else you can think of. We catch any IO errors in the Failed constructor.

Constructors

Failed

Fields

path :: !(Path Abs)

err :: IOError

Dir

Fields

path :: !(Path Abs)

fvar :: a

RegFile

Fields

path :: !(Path Abs)

fvar :: a

SymLink

Fields

path :: !(Path Abs)

fvar :: a

sdest :: File a

symlink madness, we need to know where it points to

rawdest :: !ByteString

BlockDev

Fields

path :: !(Path Abs)

fvar :: a

CharDev

Fields

path :: !(Path Abs)

fvar :: a

NamedPipe

Fields

path :: !(Path Abs)

fvar :: a

Socket

Fields

path :: !(Path Abs)

fvar :: a

data FileInfo

Low-level file information.

Constructors

FileInfo

Fields

deviceID :: !DeviceID

fileID :: !FileID

fileMode :: !FileMode

linkCount :: !LinkCount

fileOwner :: !UserID

fileGroup :: !GroupID

specialDeviceID :: !DeviceID

fileSize :: !FileOffset

accessTime :: !EpochTime

modificationTime :: !EpochTime

statusChangeTime :: !EpochTime

accessTimeHiRes :: !POSIXTime

modificationTimeHiRes :: !POSIXTime

statusChangeTimeHiRes :: !POSIXTime

sfileLike :: File FileInfo -> (Bool, File FileInfo)

fileLike :: File FileInfo -> (Bool, File FileInfo)

sdir :: File FileInfo -> (Bool, File FileInfo)

pattern FileLike :: File FileInfo -> File FileInfo

Matches on any non-directory kind of files, excluding symlinks.

pattern DirList :: () => (Functor t, Foldable t) => t (File FileInfo) -> t (File FileInfo)

Matches a list of directories or symlinks pointing to directories.

pattern FileLikeList :: () => (Functor t, Foldable t) => t (File FileInfo) -> t (File FileInfo)

Matches a list of any non-directory kind of files or symlinks pointing to such.

brokenSymlink :: File FileInfo -> (Bool, File FileInfo)

fileLikeSym :: File FileInfo -> (Bool, File FileInfo)

dirSym :: File FileInfo -> (Bool, File FileInfo)

pattern FileLikeSym :: File FileInfo -> File FileInfo

Matches on symlinks pointing to file-like files only.

pattern BrokenSymlink :: File FileInfo -> File FileInfo

Matches on broken symbolic links.

pattern DirOrSym :: File FileInfo -> File FileInfo

Matches on directories or symlinks pointing to directories. If the symlink is pointing to a symlink pointing to a directory, then it will return True, but also return the first element in the symlink- chain, not the last.

pattern DirSym :: File FileInfo -> File FileInfo

Matches on symlinks pointing to directories only.

pattern FileLikeOrSym :: File FileInfo -> File FileInfo

Matches on any non-directory kind of files or symlinks pointing to such. If the symlink is pointing to a symlink pointing to such a file, then it will return True, but also return the first element in the symlink- chain, not the last.

readFile :: (Path Abs -> IO a) -> Path Abs -> IO (File a)

Reads a file or directory Path into an AnchoredFile, filling the free variables via the given function.

readDirectoryContents

Arguments

:: (Path Abs -> IO a)
fills free a variable

-> Path Abs
path to read

-> IO [File a]

Get the contents of a given directory and return them as a list of AnchoredFile.

getContents :: (Path Abs -> IO a) -> File FileInfo -> IO [File a]

A variant of readDirectoryContents where the second argument is a File. If a non-directory is passed returns an empty list.

goUp :: File FileInfo -> IO (File FileInfo)

Go up one directory in the filesystem hierarchy.

goUp' :: Path Abs -> IO (File FileInfo)

Go up one directory in the filesystem hierarchy.

anyFailed :: [File a] -> Bool

True if any Failed constructors in the tree.

successful :: [File a] -> Bool

True if there are no Failed constructors in the tree.

failed :: File a -> Bool

Returns true if argument is a Failed constructor.

failures :: [File a] -> [File a]

Returns a list of Failed constructors only.

comparingConstr :: File FileInfo -> File FileInfo -> Ordering

isFileC :: File a -> Bool

isDirC :: File a -> Bool

isSymC :: File a -> Bool

isBlockC :: File a -> Bool

isCharC :: File a -> Bool

isNamedC :: File a -> Bool

isSocketC :: File a -> Bool

getFileInfo :: Path Abs -> IO FileInfo

Gets all file information.

handleDT :: Path Abs -> IO (File a) -> IO (File a)

rethrowFailed :: IO (File a) -> IO (File a)

Carries out the action. If the action returns a file with a failed constructor, rethrows the IOError within. Otherwise, returns the file unchanged.

isBrokenSymlink :: File FileInfo -> Bool

Checks if a symlink is broken by examining the constructor of the symlink destination.

When called on a non-symlink, returns False.

packModTime :: File FileInfo -> String

Pack the modification time into a string.

packAccessTime :: File FileInfo -> String

Pack the modification time into a string.

epochToString :: EpochTime -> String

packPermissions :: File FileInfo -> String

Pack the permissions into a string, similar to what "ls -l" does.

packFileType :: File a -> String

packLinkDestination :: File a -> Maybe ByteString

fromFreeVar :: Default d => (a -> d) -> File a -> d

Apply a function on the free variable. If there is no free variable for the given constructor the value from the Default class is used.

getFPasStr :: File a -> String

getFreeVar :: File a -> Maybe a

Gets the free variable. Returns Nothing if the constructor is of Failed.