HSFM.GUI.Gtk.Callbacks

Documentation

setGUICallbacks :: MyGUI -> IO ()

Set callbacks for the whole gui, on hotkeys, events and stuff.

setViewCallbacks :: MyGUI -> MyView -> IO ()

Set callbacks specific to a given view, on hotkeys, events and stuff.

openTerminalHere :: MyView -> IO ProcessID

closeTab :: MyGUI -> MyView -> IO ()

Closes the current tab, but only if there is more than one tab.

newTabHere :: MyGUI -> Item -> IO ()

del :: [Item] -> MyGUI -> MyView -> IO ()

Supposed to be used with withRows. Deletes a file or directory.

moveInit :: [Item] -> MyGUI -> MyView -> IO ()

Initializes a file move operation.

copyInit :: [Item] -> MyGUI -> MyView -> IO ()

Supposed to be used with withRows. Initializes a file copy operation.

operationFinal :: MyGUI -> MyView -> Maybe Item -> IO ()

Finalizes a file operation, such as copy or move.

newFile :: MyGUI -> MyView -> IO ()

Create a new file.

newDir :: MyGUI -> MyView -> IO ()

Create a new directory.

renameF :: [Item] -> MyGUI -> MyView -> IO ()

urlGoTo :: MyGUI -> MyView -> IO ()

Go to the url given at the urlBar and visualize it in the given treeView.

If the url is invalid, does nothing.

goHome :: MyGUI -> MyView -> IO ()

execute :: [Item] -> MyGUI -> MyView -> IO ()

Execute a given file.

open :: [Item] -> MyGUI -> MyView -> IO ()

Supposed to be used with withRows. Opens a file or directory.

upDir :: MyGUI -> MyView -> IO ()

Go up one directory and visualize it in the treeView.

goHistoryPrev :: MyGUI -> MyView -> IO ()

Go "back" in the history.

goHistoryNext :: MyGUI -> MyView -> IO ()

Go "forth" in the history.