HSFM.GUI.Gtk.Dialogs

Documentation

showErrorDialog :: String -> IO ()

Pops up an error Dialog with the given String.

showConfirmationDialog :: String -> IO Bool

Asks the user for confirmation and returns True/False.

fileCollisionDialog :: ByteString -> IO (Maybe FCollisonMode)

renameDialog :: ByteString -> IO (Maybe FCollisonMode)

showAboutDialog :: IO ()

Shows the about dialog from the help menu.

withConfirmationDialog :: String -> IO () -> IO ()

Carry out an IO action with a confirmation dialog. If the user presses No, then do nothing.

withErrorDialog :: IO a -> IO ()

Execute the given IO action. If the action throws exceptions, visualize them via showErrorDialog.

textInputDialog

Arguments

:: GlibString string

=> string
window title

-> string
initial text in input widget

-> IO (Maybe String)

Asks the user which directory copy mode he wants via dialog popup and returns DirCopyMode.

showFilePropertyDialog :: [Item] -> MyGUI -> MyView -> IO ()