HSFM.GUI.Gtk.MyView

Documentation

newTab :: MyGUI -> IO FMView -> Item -> Int -> IO MyView

Creates a new tab with its own view and refreshes the view.

createMyView :: MyGUI -> IO FMView -> IO MyView

Constructs the initial MyView object with a few dummy models. It also initializes the callbacks.

switchView :: MyGUI -> MyView -> IO FMView -> IO ()

Switch the existing view in MyView with the one that the io action returns.

destroyView :: MyGUI -> MyView -> IO Int

Destroys the given view by disconnecting the watcher and destroying the active FMView container.

Everything that needs to be done in order to forget about a view needs to be done here.

Returns the page in the tab list this view corresponds to.

createIconView :: IO FMView

Createss an IconView.

createTreeView :: IO FMView

Creates a TreeView.

refreshView :: MyGUI -> MyView -> Item -> IO ()

Refreshes the View based on the given directory.

Throws:

userError on inappropriate type

constructView :: MyGUI -> MyView -> IO ()

Constructs the visible View with the current underlying mutable models, which are retrieved from MyGUI.

This sort of merges the components mygui and myview and fires up the actual models.