HSFM.GUI.Gtk.Plugins

Documentation

insertPos :: Int

Where to start inserting plugins.

myplugins :: [(IO MenuItem, [Item] -> MyGUI -> MyView -> IO Bool, [Item] -> MyGUI -> MyView -> IO ())]

A list of plugins to add to the right-click menu at position insertPos.

The left part of the triple is the menuitem. The middle part of the triple is a filter function that decides whether the item is shown. The right part of the triple is the callback, which is invoked when the menu item is clicked.

Plugins are added in order of this list.

diffItem :: IO MenuItem

diffFilter :: [Item] -> MyGUI -> MyView -> IO Bool

diffCallback :: [Item] -> MyGUI -> MyView -> IO ()