HSFM.GUI.Gtk.Utils

Documentation

getSelectedTreePaths :: MyGUI -> MyView -> IO [TreePath]

getSelectedItems :: MyGUI -> MyView -> IO [Item]

Gets the currently selected item of the treeView, if any.

withItems

Arguments

:: MyGUI

-> MyView

-> ([Item] -> MyGUI -> MyView -> IO ())
action to carry out

-> IO ()

Carry out an action on the currently selected item.

If there is no item selected, does nothing.

fileListStore

Arguments

:: Item
current dir

-> MyView

-> IO (ListStore Item)

Create the ListStore of files/directories from the current directory. This is the function which maps the Data.DirTree data structures into the GTK+ data structures.

getFirstItem :: MyView -> IO Item

Currently unsafe. This is used to obtain any item, which will fail if there is none.

getCurrentDir :: MyView -> IO Item

Reads the current directory from MyView.

This reads the MVar and may block the main thread if it's empty.

pushStatusBar :: MyGUI -> String -> IO (ContextId, MessageId)

Push a message to the status bar.

popStatusbar :: MyGUI -> IO ()

Pop a message from the status bar.

rawPathToIter :: MyView -> TreePath -> IO (Maybe TreeIter)

Turn a path on the rawModel into a path that we can use at the outermost model layer.

rawPathToItem :: MyView -> TreePath -> IO (Maybe Item)

Turn a path on the rawModel into the corresponding item that we can use at the outermost model layer.

addHistory :: Eq a => a -> [a] -> [a]

Makes sure the list is max 5. This is probably not very efficient but we don't care, since it's a small list anyway.