HSFM.GUI.Gtk.Callbacks.Utils

Documentation

doFileOperation :: FileOperation -> IO ()

Carries out a file operation with the appropriate error handling allowing the user to react to various exceptions with further input.

_doFileOperation :: [Path b1] -> Path Abs -> (Path b1 -> Path Abs -> IO b) -> (Path b1 -> Path Abs -> IO a) -> IO () -> IO ()

goDir :: MyGUI -> MyView -> Item -> IO ()

Helper that is invoked for any directory change operations.