HSFM.GUI.Gtk.Data

Documentation

data MyGUI

Monolithic object passed to various GUI functions in order to keep the API stable and not alter the parameters too much. This only holds GUI widgets that are needed to be read during runtime.

Constructors

MkMyGUI

Fields

rootWin :: !Window

main Window

menubar :: !MenuBar

statusBar :: !Statusbar

clearStatusBar :: !Button

notebook :: !Notebook

fprop :: !FilePropertyGrid

settings :: !(TVar FMSettings)

operationBuffer :: !(TVar FileOperation)

data MyView

This describes the contents of the current view and is separated from MyGUI, because we might want to have multiple views.

Constructors

MkMyView

Fields

view :: !(TVar FMView)

cwd :: !(MVar Item)

rawModel :: !(TVar (ListStore Item))

sortedModel :: !(TVar (TypedTreeModelSort Item))

filteredModel :: !(TVar (TypedTreeModelFilter Item))

inotify :: !(MVar INotify)

history :: !(TVar ([Path Abs], [Path Abs]))

scroll :: !ScrolledWindow

viewBox :: !Box

rcmenu :: !RightClickMenu

upViewB :: !Button

homeViewB :: !Button

refreshViewB :: !Button

urlBar :: !Entry

data MenuBar

Constructors

MkMenuBar

Fields

menubarFileQuit :: !ImageMenuItem

menubarHelpAbout :: !ImageMenuItem

data RightClickMenu

Constructors

MkRightClickMenu

Fields

rcMenu :: !Menu

rcFileOpen :: !ImageMenuItem

rcFileExecute :: !ImageMenuItem

rcFileNewRegFile :: !ImageMenuItem

rcFileNewDir :: !ImageMenuItem

rcFileNewTab :: !ImageMenuItem

rcFileNewTerm :: !ImageMenuItem

rcFileCut :: !ImageMenuItem

rcFileCopy :: !ImageMenuItem

rcFileRename :: !ImageMenuItem

rcFilePaste :: !ImageMenuItem

rcFileDelete :: !ImageMenuItem

rcFileProperty :: !ImageMenuItem

rcFileIconView :: !ImageMenuItem

rcFileTreeView :: !ImageMenuItem

data FilePropertyGrid

Constructors

MkFilePropertyGrid

Fields

fpropGrid :: !Grid

fpropFnEntry :: !Entry

fpropLocEntry :: !Entry

fpropTsEntry :: !Entry

fpropModEntry :: !Entry

fpropAcEntry :: !Entry

fpropFTEntry :: !Entry

fpropPermEntry :: !Entry

fpropLDEntry :: !Entry

data FMSettings

FM-wide settings.

Constructors

MkFMSettings

Fields

showHidden :: !Bool

isLazy :: !Bool

iconSize :: !Int

data FMView

Constructors

FMTreeView !TreeView

FMIconView !IconView

type Item = File FileInfo

fmViewToContainer :: FMView -> Container